# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/messages": {
      "get": {
        "operationId": "MessageController_getMessages",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": false,
            "in": "query",
            "description": "Filter by chat ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Filter by sender. A phone also matches group messages via the author field and any lid that resolves to it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max messages to return (default 50)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message history"
          }
        },
        "summary": "Get message history for a session",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-text": {
      "post": {
        "operationId": "MessageController_sendText",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTextMessageDto"
              },
              "examples": {
                "minimal": {
                  "summary": "Plain text message",
                  "value": {
                    "chatId": "628123456789@c.us",
                    "text": "Hello from OpenWA!"
                  }
                },
                "withMentions": {
                  "summary": "Group message with an @mention (the text must carry the @<number> token)",
                  "value": {
                    "chatId": "120363000000000000@g.us",
                    "text": "Hello @62811",
                    "mentions": [
                      "62811@c.us"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Send a text message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-template": {
      "post": {
        "operationId": "MessageController_sendTemplate",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTemplateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template rendered and sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          },
          "404": {
            "description": "Session or template not found"
          }
        },
        "summary": "Render a stored text template and send it as a text message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-image": {
      "post": {
        "operationId": "MessageController_sendImage",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              },
              "examples": {
                "fromUrl": {
                  "summary": "Fetch the media from a URL",
                  "value": {
                    "chatId": "628123456789@c.us",
                    "url": "https://example.com/image.jpg",
                    "caption": "Check out this image!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send an image message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-video": {
      "post": {
        "operationId": "MessageController_sendVideo",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              },
              "examples": {
                "fromUrl": {
                  "summary": "Fetch the media from a URL",
                  "value": {
                    "chatId": "628123456789@c.us",
                    "url": "https://example.com/video.mp4"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send a video message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-audio": {
      "post": {
        "operationId": "MessageController_sendAudio",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendAudioMessageDto"
              },
              "examples": {
                "fromUrl": {
                  "summary": "Fetch the media from a URL",
                  "value": {
                    "chatId": "628123456789@c.us",
                    "url": "https://example.com/audio.ogg",
                    "ptt": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Audio sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send an audio/voice message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/sessions/{sessionId}/messages/send-document": {
      "post": {
        "operationId": "MessageController_sendDocument",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaMessageDto"
              },
              "examples": {
                "fromUrl": {
                  "summary": "Fetch the media from a URL",
                  "value": {
                    "chatId": "628123456789@c.us",
                    "url": "https://example.com/report.pdf",
                    "filename": "report.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Send a document/file",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.14.2",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "This instance (the origin serving these docs)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Another instance (set host and port)",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "SendTemplateMessageDto": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "description": "WhatsApp chat ID (phone@c.us for individual, groupId@g.us for groups)",
            "example": "628123456789@c.us"
          },
          "templateId": {
            "type": "string",
            "description": "Template ID to render. Provide either templateId or templateName.",
            "example": "b1c2d3e4-f5a6-7890-bcde-f01234567890"
          },
          "templateName": {
            "type": "string",
            "description": "Template name to render. Provide either templateId or templateName.",
            "example": "order-confirmation"
          },
          "vars": {
            "type": "object",
            "description": "Variables substituted into {{placeholder}} tokens in the template",
            "example": {
              "customer": "Alice",
              "orderId": "1234"
            }
          }
        },
        "required": [
          "chatId"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}